+++ CHANGELOG.rst
@@ -45,6 +45,7 @@ Fixed
 
   Reported by Guillaume Truchot (@GuiTeK)
 * Fix orquesta st2kv to return empty string and null values. (bug fix) #4678
+* Allow the orquesta st2kv function to return default for nonexistent key. (improvement) #4678
 
 3.0.0 - April 18, 2019
 ----------------------